Update order struct

+++ search.go
@@ -89,7 +89,7 @@ type CustomOrderSearchReturn struct {
 			} `json:"calculatedTaxes"`
 			TaxRules []struct {
 				TaxRate    int           `json:"taxRate"`
-				Percentage int           `json:"percentage"`
+				Percentage float64       `json:"percentage"`
 				Extensions []interface{} `json:"extensions"`
 				ApiAlias   string        `json:"apiAlias"`
 			} `json:"taxRules"`
@@ -116,7 +116,7 @@ type CustomOrderSearchReturn struct {
 			} `json:"calculatedTaxes"`
 			TaxRules []struct {
 				TaxRate    int           `json:"taxRate"`
-				Percentage int           `json:"percentage"`
+				Percentage float64       `json:"percentage"`
 				Extensions []interface{} `json:"extensions"`
 				ApiAlias   string        `json:"apiAlias"`
 			} `json:"taxRules"`
@@ -398,7 +398,7 @@ type CustomOrderSearchReturn struct {
 				} `json:"calculatedTaxes"`
 				TaxRules []struct {
 					TaxRate    int           `json:"taxRate"`
-					Percentage int           `json:"percentage"`
+					Percentage float64       `json:"percentage"`
 					Extensions []interface{} `json:"extensions"`
 					ApiAlias   string        `json:"apiAlias"`
 				} `json:"taxRules"`
@@ -561,7 +561,7 @@ type CustomOrderSearchReturn struct {
 				} `json:"calculatedTaxes"`
 				TaxRules []struct {
 					TaxRate    int           `json:"taxRate"`
-					Percentage int           `json:"percentage"`
+					Percentage float64       `json:"percentage"`
 					Extensions []interface{} `json:"extensions"`
 					ApiAlias   string        `json:"apiAlias"`
 				} `json:"taxRules"`
@@ -574,7 +574,7 @@ type CustomOrderSearchReturn struct {
 				Price    float64 `json:"price"`
 				TaxRules []struct {
 					TaxRate    int           `json:"taxRate"`
-					Percentage int           `json:"percentage"`
+					Percentage float64       `json:"percentage"`
 					Extensions []interface{} `json:"extensions"`
 					ApiAlias   string        `json:"apiAlias"`
 				} `json:"taxRules"`
@@ -651,7 +651,7 @@ type CustomOrderSearchReturn struct {
 				} `json:"calculatedTaxes"`
 				TaxRules []struct {
 					TaxRate    int           `json:"taxRate"`
-					Percentage int           `json:"percentage"`
+					Percentage float64       `json:"percentage"`
 					Extensions []interface{} `json:"extensions"`
 					ApiAlias   string        `json:"apiAlias"`
 				} `json:"taxRules"`
